src/account/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
cryptography
pymysql
mysql-connector-python
requests
PyJWT
cryptography